Add an InputClassifier HAL
The new HAL will be responsible for processing touch video frames that
are received from InputClassifier stage.
After processing, the HAL will generate a classification of the current
gesture. This classification will later be provided to the app via
the MotionEvent.getClassification() public API.

The eventual goal is to classify each event stream as a force touch
or not a force touch.

An example implementation of the HAL is presented here. This code was
used for local testing on blueline, and should not be actually installed
or used in any way other than for reference.

Bug: 62940136
Test: observed "detected deep press" logs when pressing with large
finger area on the touchscreen
Change-Id: Id6ac4337435e4ac07877da11ca184b6dd4d64780