Add Emergency Call Routing in HAL
Some countries or carriers require some emergency numbers that must
be handled with normal call routing or emergency routing.

In multi-sim senario, this radio request will be sent through the IRadio
service that serves the subscription the emergency number belongs to,
no matter of the PUK/PIN state of the subscription and the service state.

Test: Treehugger
Bug: 112657134
Change-Id: Iaa9768226dc2d7d2d66a9678823ba7d0047a1988