Add a new parameter to emergencyDial().
The new parameter is to indicate if the request is initiated
by the user through an emergency dialer or shortcut, meaning
that the request is explicitly an emergency dial request.
This is to indicate to the modem that no disambiguation (regarding
whether this is really an emergency request) should
be done for this request.

Test: builds fine
Bug: 123101600
Merged-in: Ic830762d8a8319f494f22d875ca6adad91ffce3b
Change-Id: Ic830762d8a8319f494f22d875ca6adad91ffce3b
(cherry picked from commit 143d32860fbe08eef95f582b3e9b5b65762f2706)