Explicitly Document INT_MAX for Invalid Fields
When fields are inapplicable, they should not
be set to a clearly out-of-range value to signal
to the framework that those fields are not used.
In some cases, there is an in-range invalid value
that has been defined by the standards. The docs
are inconsistent in calling out INT_MAX as the
value to be used when something is inapplicable
vs the case when a measurement is simply not
reported. In all cases, INT_MAX can be used to
denote an invalid value, and in cases where the
field/structure is inapplicable, it is the correct
value. This CL updates all the docstrings for
SignalStrength-related fields to clarify that
INT_MAX is the correct "invalid" value for cases
when fields are inapplicable.

Bug: 123088652
Test: compilation (docstring-only change); CTS
    naturally enforces this change on devices with
    newer HALs; backwards compatibility is preserved
    for existing HAL versions.

Change-Id: I5cfa917f504d15691ab3f2c298189bdd47794a42