Document that video frames rotate
We recently changed the behaviour in InputReader to adjust the video
frame data based on the current orientation.
Update the HAL documentation to reflect that.

Also, rearrange the order of width and height, because it is more
natural to have the format of num of rows x num of columns. So
currently, it is reversed, and not as natural.
More information about this is in ag/6414079.

Bug: 123241238
Test: none
Change-Id: I69f184794bd7684e5c0981a6db66542ab914eff2