compatibility_matrix: restore P graphics.composer
graphics.composer 2.3 was added in Android Q, so it is in the
compatibility matrix for target-level 4. Devices updating from P
to Q will automatically have this option applies. However, for
sanity, we try to keep the P compatibility matrix between releases.

This change should have no functional difference except minimizing
the delta in this file between P/Q branches.

Bug: 126892442
Test: diff compatibility matrix files
Test: these matrices are verified at build time
Change-Id: I8514c7b6ce4c2670ce7926fb31b09ed783d17dca