Return display capability if getDisplayBrightnessSupport is not registered.
If getDisplayBrightnessSupport is not registered, there are two possibilities,
one is it's not supported at all, the other is that the support is returned in
getDisplayCapabilities. And thus we need to check getDisplayCapabilities, and
we return UNSUPPORTED always in this case. This patch also allows
getPerFrameMetadataKeys to return UNSUPPORTED on non-HDR capable devices.

BUG: 131595097
Test: Build, boot.
Change-Id: Ied302b1ac702dd94e039f1081d5420395c1bfbf4