Fix so that dual access points (AP) works.
Devices that only have dual AP setup without an STA were not working. A
use case for this would be in a car, having a wifi hotspot active while
also projecting to the car head unit from your phone. This was working
but a recent change in CL a3e5b7f that forced STA to be on wlan0 and AP to start at
wlan1 broke this functionality. This change is to augment the original
CL to keep it's function in place while also taking into
account devices that have dual AP.

Bug: 132268977
Test: Manual boot of hawk device along with testing of tethering and
local only hotspot.

Change-Id: I8da5dd4f0baea1bf0f22d3275e356936a58fb3d1