Fix wait_for timestamps in Sensors VTS
Previously, NoStaleEvents was treating any timestamps it dealt with as
if they were in microseconds, but sensors.minDelay is in microseconds
and Event timestamps are in nanoseconds. This uses std::chrono helpers
to ensure the correct time is used when deciding how long to sleep
during the test so that if waitForEvents never passes, the test doesn't
time out.

Bug: 136736906
Test: Run VTS and verify VtsHalSensorsV2_0Target doesn't finish as an
    incomplete module.

Change-Id: Ibba59dbf9312f97d7275e5aa8cd36547ab09e328
Merged-In: Ibba59dbf9312f97d7275e5aa8cd36547ab09e328