Set up shell to use for unit tests
Sets up a unit test file that can be used to unit test the HalProxy. The
unit tests will eventually end up containing code that injects the fake
SensorsSubHal implementation into the HalProxy and verifies various
parts of the implementation. This will likely require some modifications
to the SensorsSubHal class to modify it during test execution, but it's
better that the same implementation is shared between unit / manual
integration testing.

Test: Run atest android.hardware.sensors@2.0-halproxy-unit-tests
Bug: 136511617
Change-Id: I7b6865564ea41ab3c58f77d3168366c95e3289e5