Fix RadioHidlTest_v1_2.startNetworkScan maxSearchTime
Add the initialization of the incrementalResultsPeriodicity and
maxSearchTime in the networkScan request.
The incrementalResultsPeriodicity and maxSearchTime are not
initialized in the networkScan request. So the default value is 0.
But the incrementalResultsPeriodicity and maxSearchTime
are required not to be 0 in subsequent tests.

CL ported from pag/1433565

Bug:140451434
Test: make vts && vts-tradefed run commandAndExit vts -m \
      VtsHalRadioV1_2Target -t RadioHidlTest_v1_2#startNetworkScan
Change-Id: I1b4e512a37d3d85ebc905a64987a40af801bce53
(cherry picked from commit f7894eb4d8c395add7c70a20520336057ae2656f)