Fix the failed case for the VtsHalWifiSupplicantV1_2Host#SupplicantStaIfaceHidlTest.RegisterCallback_1_2
From the log and case logic,when assertTrue isSupplicantRunning is not the real wpa_supplicant status.
So occasionally there will be a timeout when waiting for a callback in the waitForHidlService method.
Because this is a low memory version and it is gsi image, we can't modify the image, so we can solve it by extending the time.

Bug:[142585026]
Test:run vts of the VtsHalWifiSupplicantV1_2Host.
Change-Id: Ideef032d310165d956c3ca0967b26bc3e3f44c3b