DO NOT MERGE: Audio HAL: do not test input stream if no Built-in mic on primary
The test used to always test input stream, assuming that all devices had
built-in device on the primary Module. Nevertheless, although uncommon,
the mic could be on any module or even not exist.

This patch makes sure that the input stream tests are only run if there
is a Built-in mic on the primary module.

This patch also fixes GetMicrophonesTest to accept NOT_SUPPORTED
result.

This patch is specific for Android P. Later versions already have
these fixes.

Bug: 114303641
Test: atest VtsHalAudioV4_0TargetTest
      on device with a built-in mic and on a device w/o

Change-Id: I7289724e5a73c1ffd09ca990f681844bdc8f6b3e