Fix a resource clean-up routine
When a test case is being terminated, VTS framework calls TearDown() to
clean up all resources.  VtsHalEvsV1_1TargetTest stores weak references
to opened hardware cameras for this and tries to close them explicitly
when it can promote them.  However, it was observed that VirtualCamera's
destructor is not called yet even it fails to promote.

Therefore, this change modifies a container to hold a strong pointer and
explictly removes it after a test case calls closeCamera().

Also, this change corrects a bug in checking setExtendedInfo_1_1()
results.

Bug: 150893461
Test: vts-trafed VtsHalEvsV1_1_TargetTest
Change-Id: I23b55b2f2282834fb5a9d364f4eb176060027de1