remove stopAllRecognitions VTS test
The stopAllRecognitions API is not used at the layers above the HAL, and
it requires the HAL to own the state of all active recognitions. This
API causes discontinuity with the upper layers also maintaining state
information.

The VTS requirement is being removed.

Bug: 151281377
Test: atest VtsHalSoundtriggerV2_0TargetTest
&& atest VtsHalSoundtriggerV2_1TargetTest

Change-Id: If14ed8177e0ff2730d2ed78ba8fbbbd058c4d57a
Merged-In: If14ed8177e0ff2730d2ed78ba8fbbbd058c4d57a