Radio tests: root before multi-sim preparation.
These tests assumed that they were rooted before running, and due to
a combination of b/152655658 and b/154638140, these appeared to work
when a device already happened to be rooted and the test failures showed
up as problems with the test rather than as problems with the target
preparer.

This is the same workaround being used elsewhere.

Fixes: 154445273 # specifically about VtsHalSapV1_0TargetTest
Test: atest VtsHalSapV1_0TargetTest # passes 100%

The following tests, I confirm MultiSimTargetPreparer works, although my
test device has no SIM so I cannot get fully useful results.
Test: atest VtsHalRadioV1_0TargetTest # slot2 cases can now pass
Test: atest VtsHalRadioV1_1TargetTest # slot2 cases can now pass
Test: atest VtsHalRadioV1_2TargetTest # slot2 cases can now pass
Test: atest VtsHalRadioV1_3TargetTest # fully passes
Test: atest VtsHalRadioV1_4TargetTest # fully passes

Change-Id: Id62cb2e1b7a95552b560ab1d4ff2f63c0a7569a3