Reduce the number of iterations in PreEnrollTest
The current number of iterations is too high and results in a timeout
from the testing framework.

The goal of the test is to make sure that the generated challenges are
different on every run. It's impossible to test whether any two
challenges will ever be the same for any number of iterations. Hence,
since a 1000 iterations doesn't provide much more value over 10
iterations, 10 was chosen because of its better performance.

Bug: 155844206
Test: "atest VtsHalBiometricsFingerprintV2_1TargetTest" on a blueline
Change-Id: I5db887bdcfed8fb0e44fcedcf9e99dd185d395c3