Avoid unnecessary access of BoringSSL structs.
Checking cert_info->key->algor->algorithm is redundant with the checks
following it. If the public key is an EC key, that was the OID. Remove
the check so this code does not break when BoringSSL makes the X509
structures opaque in the future.

While we're not particularly aiming to make ECDSA_SIG opaque, getters
exist, so go ahead and use them.

Test: mm, treehugger
Change-Id: I1b37fef2290b7697a6e821f20ba702b3da5ef18d