Add API to get the shared memory Audio/Video handle
In the MediaEvent of AV filters, the AV memory handle is mapped/unmapped on each event.
If the vendor allocates a persistent memory block, and use MediaEvent.offset to access the data,
the mapping change is unnecessary.

To improve this, a new API to get shared AV Memory is introduced.
To pass additional information to Codec2 for AV memory,
the index for Codec2 to setup additional information can be carried in the handle.

Test: atest VtsHalTvTunerV1_1TargetTest
Bug: 162013047
Bug: 157143515
Bug: 156664393
Change-Id: Ibd4471ff8e19b0a99b8bcbc982dc08dbcc92efcf