health: Add CAP_BLOCK_SUSPEND
Add CAP_BLOCK_SUSPEND to android.hardware.health@1.0 and
android.hardware.health@2.1 service so that when epoll_wait()
is waiting on uevents and timer_fd events it can block suspend
to take the necessary action. Without this permission, eventpoll
is going to ignore the flags EPOLLWAKEUP used in health HAL or
healthloop.

Bug: 169873088
Change-Id: If9b0e1f60e47d7bf54942f589acbfbb883b9bac8