wifi: Add sleep 3 seconds to allow terminate processing in VTS tear down
In current android design, the general use case is the interface is
removed after terminate() call.
But the VTS kill the hostapd immediately it causes terminate doesn't
work normally.
Add sleep 3 seconds to allow terminate processing.

Another test case is considered to check hostapd died because any reason.
(for instance: NPE crash)
But the interface handling in android design is control in framework and WiFiHal.
Case 1. Native crash will be handled in framework and clean the interface for
the recovery.

Case 2. VTS use "stopHAL" to clean "primy interface".
It will cause clean up fail if OEM use "non-primy interface" for the SoftAp.
But it is not requirement in hostapd HAL. Verify terminate is good in
VtsHalWifiHostapd.

Bug: 163170239
Test: atest -c VtsHalWifiHostapdV1_1TargetTest
Test: atest -c VtsHalWifiHostapdV1_0TargetTest
Change-Id: I25618961eb43492a10a293484c9af431d5971e1b