wifi: Support use "non-primy interface when testing hostapd Hal
In current android design, the general use case is the interface is
removed after terminate() call.
But the VTS kill the hostapd immediately it causes terminate doesn't
work normally.
Add sleep 3 seconds to allow terminate processing.

Another test case is considered to check hostapd died because any reason.
(for instance: NPE crash)
But the interface handling in android design is control in framework and WiFiHal.
Case 1. Native crash will be handled in framework and clean the interface for
the recovery.

Case 2. VTS use "stopHAL" to clean "primy interface".
It will cause clean up fail if OEM use "non-primy interface" for the SoftAp.
But it is not requirement in hostapd HAL. Verify terminate is good in
VtsHalWifiHostapd.

Bug: 177483254
Bug: 177389456
Test: atest -c VtsHalWifiHostapdV1_3TargetTest
Change-Id: I39f6ee92b1354ab2cebbe2fe6e9873e62d850c42