STHAL 2.4
This version of the soundtrigger HAL adds support for dynamic resource
management by the HAL, allowing it to refuse to start/load and to
preemptively stop/unload models at its discretion.

This give the HAL more freedom to make resource allocation decisions
and avoids the need to precisely model the various constraints and
have them enforced by the framework.

Bug: 178722883
Change-Id: I58807407417fc099b7538e685309d2c846ed66fd