Add MPEG-H enum.
MPEG-H is a new surround audio codec. Add new enum format for MPEG-H in
the HIDL interface. Though MPEG-H has different profiles and audio
levels, only low complexity and baseline with audio level as 3 and 4 is
defined in this CL as they are relevant for the current and upcoming
commercial services.

Bug: 178619392
Test: make
Change-Id: I5672f3874f6cffe31a83b6d7e351d0bf56bfd038