audio: Update audioPolicyConfiguration version
HAL V7 uses spaces instead of ',' and '|' for delimiting
elements in channel masks, sample rates, and flags lists
in the Audio Policy Manager configuration XML.
The version specified in the APM XML has been updated
from '1.0' to '7.0' in order to reflect this.

The migration script has been updated to change the
version in the XML on update.

Bug: 142480271
Test: m
Change-Id: I92c206dddee6964fc2019fed80435cb8d69dbfe9