Tweak ContextHub HAL 1.2 based on feedback
To assist in making permissions attribution more robust at the
ContextHub service level, tweak the V1.2 interface with the following
changes:
- Return the full list of supported permissions that all hubs support
for attribution. This assists in limiting the number of permissions
strings sent between the HAL and service to only those the HAL actually
attempts to use.
- Add new parameter to handleClientMsg_1_2 that allows nanoapps to
denote the list of attributable permissions relating to the contents of
the message being sent. This is strictly a subset of the permissions the
nanoapp holds and allows a message to be sent without attributing the
full nanoapp permission set to the host client when the message may not
have been generated from data covered by the permissions.

Bug: 166846988
Test: compile
Change-Id: I2dc8e1ab4fce2a9ebcc393d07fdffe23dfceb4c2