Fix IFingerprint VTS
The main function was not defined/visible because it was inside of a
namespace. The default GTest main was used instead, and the Binder
threadpool was never initialized, causing the test to hang.

In addition to that, the test expected a single onStateChanged with
SessionState::AUTHENTICATING. Added support and expectations for
multiple callback invocations.

Bug: 181184674
Test: atest VtsHalBiometricsFingerprintTargetTest
Change-Id: I1da72a81b091c7382494a2c4828c113671de845c