Fix RefBase Explicit destruction
For an object to be managed by std smart pointer, its base class should NOT be "RefBase".  It is not only unnecessary but also risking the object being owned by android smart pointer as well, which may destroy the object two times.

This issue may be detected by the destructor of class RefBase, with the log complaint "RefBase: Explicit destruction,..."

Test: monkey test for one day and one night

Signed-off-by: Jintao Zhu <zhujtcsieee@gmail.com>
Change-Id: If7736770983e17f5a4125b19dfbd1745ee5a9589