Fix SetCountryCode test when US is not supported
SupplicantStaIfaceHidlTest.SetCountryCode(default,default)_32bit fails
when US countrycode is not supported on the device. The test will
read the value of "ro.boot.wificountrycode" and set it to the
countrycode. When the property is not defined on the device, US will
be set by default.

Bug: SupplicantStaIfaceHidlTest.SetCountryCode(default,default)_32bit fails
Bug id https://issuetracker.google.com/182296217
Test: Module VtsHalWifiSupplicantV1_0Host passes successfully

Change-Id: Iec2991223c6857d65b2ecc4337e920d5b823dd3d
Signed-off-by: Hassen KETATNI <hassen.ketatni@sagemcom.com>