[SP37] Change requirement of older event callback.
Currently, vendors are asked to duplicate old events to onEvent()
and onEvent_1_1(). This is not the typically way other modules
would do. Thus, remove the requirement of duplicate events,
and replace it with only firing onEvent_1_1 when feasible.

Note no VTS is added currently, from vendor feedback, there is
no reliable way to guarantee the callbacks will be called in the
test.

Test: hidl-gen -L hash -r android.hardware:hardware/interfaces \
      android.hardware.tetheroffload.control@1.1
Bug: 149467454
Change-Id: I75c1a7788e5d48fb3e2836d54b4fd854c0247214