Add default VHAL test.
Add tests for default VHAL implementation. This CL also modifies
the onDump interface for IVehicleServer. The onDump interface now
returns the dumped info to the caller and let the caller print
the info to hidl_handle. This is due to IVehicleServer might run
in non-android VM that has no notion of HIDL handle.

Test: atest

Change-Id: I4fc5dedf69d6730b1b196ce171be08339ee7fbd2
Test: run test