Remove vehicle binding seed property in emulator
Emulator does not support vehicle binding seed property,
but it reports it supports it. Remove it for now.
Otherwise the vehicle binding utilities will try to set the
vehicle binding seed but that will not be saved anywhere
in the emulator.

Test: manual
Bug: 157501579
Change-Id: I9e8ad382a49e1108e0e25d6eee94b24d766264e8
Merged-In: I874e94b06fe675a96e5b15c9bff087023b4ea109