Audio VTS: Make the active microphone query test more robust in V7
Prior to V7 the test which exercises
IStreamIn.getActiveMicrophones was using a hardcoded
configuration for the input stream. This configuration no longer
works for some of new devices. To fix that, the part of the test
which calls getActiveMicrophones has been moved into a separate
test--a descendant of InputStreamTest which is parametrized
using the actual configuration of the DUT.

Tests for HAL versions prior to V7 are not affected
because they don't use a full parser for the DUT config.

Bug: 193849687
Test: atest VtsHalAudioV7_0TargetTest
Change-Id: I00fe8fedb6bfc6e034387b35c88f954cb2638dfa
(cherry picked from commit e4ce86bfb7bd81eea39da5d76dbf796e2c1dbe1d)
Merged-In: I00fe8fedb6bfc6e034387b35c88f954cb2638dfa