Revert "Clarify the ComposerClient comment for CONFIG_GROUP"
This reverts commit 37c5495c52f200a26ea4ac064180af7f2f8ae722.

This "clarification" was in fact a change of the HAL interface,
introducing a new requirement which was not present in earlier
Android releases. There are devices on the market that switch
seamlessly between configs that differ in resolution and DPI,
and therefore have those configs in a config group.

Bug: 200184776
Merged-In: I5bba1c6532f5f46da874385164a0bc0cc0435225
Change-Id: I5bba1c6532f5f46da874385164a0bc0cc0435225