Bluetooth HAL: Add missing retry logic to the H4Protocol::Send()
As part of an optimization, H4Protocol::Send() is modified to
use writev() instead of WriteSafely(), which has a retry logic
to send remaining bytes if the underlying write() call couldn't
transfer all the bytes in a single system call. This cl is now
adding this missing retry logic to this function.

Bug: 198992020
Test: H4 UTTest, Bluetooth functional test.

Change-Id: I78fad886d52f0910c83b000d5768128f00d3ddbc