Revert "VTS: Test that configs in a group differ only by vsync period"
This reverts commit 6f36dd6e31ac1c2d62cd9283a4d07c7c684e5872.

Differing only by vsync period was not a requirement on Android R, and
it shouldn't be added after-the-fact to an existing HAL version.

There are devices on the market that switch seamlessly between configs
that differ in resolution/dpi, and therefore have those configs in a
common group.

Bug: 200184776
Merged-In: Id0bfc67e55d5139fddb2b359cabafd9281c33734
Change-Id: Ib6e6a8c5f45c9829ea30d031e1cfa9a4fb775535