send <request active source> command to test cec message
VtsHalTvCec uses <give physical address> command with broadcast
mode to test cec sendMessage function, however according to the
cec spec about the command's description which says "A request to
a device to return its Physical Address", so this command is not
right for broadcast, some cec drivers deny to broadcast this
command, which will obviously lead the VTS case fail.

So I suggest to use <request active source> command to do the job.
I have verified the case can pass on Synaptics platform with this patch.

bug-id: b/206924358

Change-Id: I040c72180cf90cfbceb6363a50a621acf888c654