Convert InputClassifier HAL to AIDL
No changes to the HAL are made in this change, we are only updating the
interface to use AIDL.
At the same time, the HAL is getting renamed to 'InputProcessor', which
will mean that it's more general-purpose than just classification. We
will be adding more stuff to this HAL in the future.

Bug: 167946763
Bug: 205761620
Test: integration tested with other framework changes on top
Change-Id: I8046cd5c74e1b69f85f21b28d5c4c3b88a61bb33