Remove 'ops' field from BCC public keys.
The application of ops is currently inconsistent within the spec, and
wrong where it is specified. It is mapped to an integer when it should
be mapped to an array that contains the integer value. Given that all
key usage is strictly defined within the spec itself, there is no value
currently in including the ops field. Because of this, the choice is
being made to simply remove the field.

Test: documentation change only
Change-Id: Ia3a9895c7840e3cda9c2775080b7dcb18544d417