Remove default value for AP_POWER_STATE_REQ.
Remove the default value for AP_POWER_STATE_REQ because initially the
device should be in an OFF state, but there is no such
AP_POWER_STATE_REQ. If we set the default value to ON, then when
VHAL set the value to ON, it would cause no value update event to
inform CarService.

Test: atest FakeVehicleHardwareTest.
Manually run audio test on cf target.
Bug: 218924258

Change-Id: I3552a38bf2433015767c2254e42871cee5c9e716