camera: Fix AIDL VTS test
The default value of a Stream's `dynamicRangeProfile` field is invalid,
which means the caller needs to set it to a sensible default with each
call. This CL updates the Stream creation in VTS test to use
`ANDROID_REQUEST_AVAILABLE_DYNAMIC_RANGE_PROFILES_MAP_STANDARD`
wherever a Stream object was being created with implicit default.

Bug: 210912368
Test: Existing VTS tests now pass on Cuttlefish
Change-Id: Iaae2f6dfde852fc9c076a6029854ccdf8f68d39a