Add -Wno-missing-permission-annotation for soundtrigger3
Soundtrigger3 has been built with "-Werror -Weverything" (from
Android.bp) and "-Weverything -Wno-missing-permission-annotation" (from
the AIDL build system). The former is added only for the ToT version of
an AIDL interface.

So far, soundtrigger3 wasn't frozen and therefore the implicit version 1
was the ToT version. As a result, absence of @RequiresPermission in its
interface didn't break the build due to the automatically added
"-Wno-missing-permission-annotation" despite that it had "-Werror".

However, this is causing build breakage as soundtrigger3 now has version
1 and 2 (ToT) and the version 1 is built *without*
  "-Wno-missing-permission-annotation".

Fixing this by manually adding the diag flag so that it's applied to all
versions.

Bug: 225941299
Test: m
Change-Id: I3a43fc77997ef900b0f173c88530f570e436bc77