DO NOT MERGE: Disable current matrix.
The current matrix should not be used in Android T. However, it is left
here so that any late fixes to the T matrix can be made in a way that
will flow downstream.

Reason for 'DO NOT MERGE' - avoid landing this change in goog/master,
since the current matrix has HALs added there, some devices may be
relying on it.

In order to allow the (AOSP subset of) the T manifest to be frozen in
AOSP, and the current matrix to be enabled from aosp/master -..->
goog/master, this will be reverted in tm-dev-plus-aosp immediately after
it merges.

Bug: 203103997
Test: boot devices and vts_treble_vintf_vendor_test passes
Change-Id: I15bacedd0820e0086a742b97ee18e097ef8086e5